pyproject.toml
[tool.poetry]
name = "zpacloud-ansible"
version = "2.1.0"
description = "Ansible collection for Zscaler Private Access (ZPA)"
authors = ["Zscaler Technology Alliances <devrel@zscaler.com>"]
license = "MIT"
keywords = ["zpacloud", "zpacloud-ansible", "zpa"]
readme = "README.md"
homepage = "https://github.com/zscaler/zpacloud-ansible"
repository = "https://github.com/zscaler/zpacloud-ansible"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9.2,<4.0"
zscaler-sdk-python = ">=1.9.0"
aiohttp = ">=3.8.4"
dpath = ">=2.1.5"
xmltodict = ">=0.12.0"
netaddr = ">=1.2.1"
pycountry = ">=23.12.11"
Babel = ">=2.14.0"
sphinx-ansible-theme = ">=0.10.3"
python-box = ">=7.3.2"
responses = ">=0.25.3"
flake8 = "*"
pydash = "*"
jinja2 = ">=3.1.0"
toml = ">=0.10.2"
pytz = ">=2024.1"

[tool.poetry.group.dev.dependencies]
black = ">=24.3.0"
pytest = "^8.2.1"
coverage = ">=4.5.4"
pylint = "2.13.*"
pytest-xdist = "*"
pytest-mock = "*"
voluptuous = "*"
yamllint = "*"
sphinx = "^7.4.7"
sphinx-rtd-theme = "*"
sphinx_ansible_theme = "*"
antsibull-docs = "^2.11.0"
antsibull-changelog = "^0.26.0"
ansible-core = "*"
ansible-lint = "*"
rstcheck = "*"
six = "*"
pycodestyle = "*"

certifi = ">=2023.11.17"
chardet = ">=5.2.0"
idna = "3.10"
requests = ">=2.22.0"
urllib3 = ">=2.0"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests/unit"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"